requirements.txt
faiss-cpu==1.7.2
deepface==0.0.73